# api/openapi.bundled.yaml
openapi: 3.2.0
...
info:
  title: Arches Platform API
  description: The Arches Platform API for managing projects and code generation
  version: v0.0.0
  summary: API for Arches platform
  termsOfService: https://archesai.com/terms
  license:
    name: AGPL-3.0
    url: https://opensource.org/licenses/AGPL-3.0
  contact:
    name: Arches AI Support
    email: support@archesai.com
    url: https://archesai.com/support
...
servers:
  - url: https://api.archesai.dev
    description: Remote development server
  - url: https://api.archesai.com
paths:
...
  /auth/verify-email:
    post:
      x-internal: auth
      operationId: ConfirmEmailVerification
      x-codegen-custom-handler: true
      summary: Confirm e-mail verification
      tags:
        - Auth
      description: This endpoint will confirm your e-mail with a token
      requestBody:
        description: Email verification request with token
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  description: The email verification token
                  type: string
                  minLength: 1
                  maxLength: 512
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                  pattern: ^[A-Za-z0-9_-]+$
              required:
                - token
        required: true
      responses:
        '200':
          $ref: '#/components/responses/EmailVerificationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /organizations:
    get:
      x-internal: auth
      operationId: ListOrganizations
      summary: List organizations
      tags:
        - Organization
      description: List organizations
      parameters:
        - $ref: '#/components/parameters/OrganizationsFilter'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/OrganizationsSort'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/OrganizationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      x-internal: auth
      operationId: CreateOrganization
      summary: Create an organization
      tags:
        - Organization
      description: Create an organization
      requestBody:
        description: Organization creation request
        content:
          application/json:
            schema:
              type: object
              properties:
                billingEmail:
                  description: The billing email to use for the organization
                  type: string
                  example: example@archesai.com
                  minLength: 1
                  maxLength: 255
                  format: email
                organizationID:
                  description: UUID identifier
                  type: string
                  format: uuid
                  example: 550e8400-e29b-41d4-a716-446655440000
                  minLength: 36
                  maxLength: 36
              required:
                - billingEmail
                - organizationID
        required: true
      security:
        - bearerAuth: []
      responses:
        '201':
          $ref: '#/components/responses/OrganizationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /organizations/{organizationID}/invitations:
    get:
      x-internal: auth
      operationId: ListInvitations
      summary: List invitations
      tags:
        - Invitation
      description: List invitations
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
        - $ref: '#/components/parameters/InvitationsFilter'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/InvitationsSort'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/InvitationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      x-internal: auth
      operationId: CreateInvitation
      summary: Create an invitation
      tags:
        - Invitation
      description: Create an invitation
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        description: Invitation creation request
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: The email of the invitated user
                  type: string
                  example: example-string
                  minLength: 1
                  maxLength: 255
                  pattern: ^[\w\s\-.,!?()@#+/]*$
                role:
                  description: The role of the invitation
                  type: string
                  enum:
                    - admin
                    - owner
                    - member
                  example: admin
              required:
                - email
                - role
        required: true
      security:
        - bearerAuth: []
      responses:
        '201':
          $ref: '#/components/responses/InvitationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    User:
      description: Schema for User entity
      title: User
      x-codegen-schema-type: entity
      x-codegen:
        repository:
          indices:
            - email
          additionalMethods:
            - name: GetUserByEmail
              params:
                - name: email
                  type: string
              returns: single
            - name: GetUserBySessionID
              params:
                - name: sessionID
                  type: string
                  format: uuid
                  maxLength: 36
              returns: single
      x-internal: auth
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          properties:
            email:
              description: The user's email address
              type: string
              format: email
              minLength: 5
              maxLength: 255
              example: user@example.com
            emailVerified:
              description: Whether the user's email has been verified
              type: boolean
              default: false
              example: true
            image:
              description: The user's avatar image URL
              type:
                - string
                - 'null'
              minLength: 5
              maxLength: 2048
              format: uri
              example: https://example.com/avatar.jpg
            name:
              description: The user's display name
              type: string
              minLength: 1
              maxLength: 255
              example: John Doe
              pattern: ^[\w\s\-.,!?()@#+/']+$
          required:
            - email
            - emailVerified
            - name
            - image
      unevaluatedProperties: false
...
  parameters:
    PageQuery:
      in: query
      name: page
      required: false
      style: form
      explode: true
      schema:
        $ref: '#/components/schemas/Page'
      description: The page parameter
      example:
        limit: 10
        offset: 0
...
    InvitationsFilter:
      in: query
      name: filter
      required: false
      style: deepObject
      explode: true
      schema:
        $ref: '#/components/schemas/FilterNode'
      description: Filter by field values
      example:
        type: eq
        field: status
        value: PENDING
    InvitationsSort:
      in: query
      name: sort
      required: false
      style: form
      explode: true
      schema:
        type: array
        maxItems: 10
        items:
          type: object
          required:
            - field
            - order
          properties:
            field:
              type: string
              enum:
                - createdAt
                - id
                - updatedAt
                - email
                - expiresAt
                - inviterID
                - organizationID
                - role
                - status
              example: createdAt
            order:
              type: string
              enum:
                - asc
                - desc
              example: asc
      description: The sort parameter
      example:
        - field: createdAt
          order: desc
...
    OrganizationID:
      schema:
        $ref: '#/components/schemas/UUID'
      in: path
      name: organizationID
      required: true
      description: The unique identifier of the organization.
    OrganizationsFilter:
      in: query
      name: filter
      required: false
      style: deepObject
      explode: true
      schema:
        $ref: '#/components/schemas/FilterNode'
      description: Filter by field values
      example:
        type: eq
        field: name
        value: Example Organization
    OrganizationsSort:
      in: query
      name: sort
      required: false
      style: form
      explode: true
      schema:
        type: array
        maxItems: 10
        items:
          type: object
          required:
            - field
            - order
          properties:
            field:
              type: string
              enum:
                - createdAt
                - id
                - updatedAt
                - billingEmail
                - credits
                - logo
                - metadata
                - name
                - plan
                - slug
                - stripeCustomerIdentifier
              example: createdAt
            order:
              type: string
              enum:
                - asc
                - desc
              example: asc
      description: The sort parameter
      example:
        - field: createdAt
          order: desc
...
  responses:
    BadRequest:
      description: 400 Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    NotFound:
      description: 404 Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
    Unauthorized:
      description: 401 Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    UnprocessableEntity:
      description: 422 Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    EmailVerificationResponse:
      description: Email verified successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              session:
                $ref: '#/components/schemas/Session'
              user:
                $ref: '#/components/schemas/User'
            required:
              - session
              - user
            additionalProperties: false
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    InvitationListResponse:
      description: Invitations retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/Invitation'
              meta:
                $ref: '#/components/schemas/PaginationMeta'
            required:
              - data
              - meta
            additionalProperties: false
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    InvitationResponse:
      description: Invitation retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Invitation'
            required:
              - data
            additionalProperties: false
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    OrganizationListResponse:
      description: Organizations retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/Organization'
              meta:
                $ref: '#/components/schemas/PaginationMeta'
            required:
              - data
              - meta
            additionalProperties: false
          example:
            data:
              - id: 550e8400-e29b-41d4-a716-446655440000
                name: Acme Corporation
                slug: acme-corp
                logo: https://example.com/org-logo.png
                billingEmail: billing@acme.com
                plan: STANDARD
                credits: 1000
                stripeCustomerIdentifier: cus_1234567890
                createdAt: '2024-01-15T09:30:00Z'
                updatedAt: '2024-01-15T09:30:00Z'
            meta:
              total: 1
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    OrganizationResponse:
      description: Organization retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Organization'
            required:
              - data
            additionalProperties: false
          example:
            data:
              id: 550e8400-e29b-41d4-a716-446655440000
              name: Acme Corporation
              slug: acme-corp
              logo: https://example.com/org-logo.png
              billingEmail: billing@acme.com
              plan: STANDARD
              credits: 1000
              stripeCustomerIdentifier: cus_1234567890
              createdAt: '2024-01-15T09:30:00Z'
              updatedAt: '2024-01-15T09:30:00Z'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
  securitySchemes:
    bearerAuth:
      description: API Token for authenticated requests
      scheme: bearer
      type: http
    sessionCookie:
      description: Session cookie for authenticated requests
      type: apiKey
      name: __Secure-better-auth.session_token
      in: cookie
security:
  - bearerAuth: []
  - sessionCookie: []