# pkg/auth/gen/controllers/confirm_email_verification.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// ConfirmEmailVerification - POST /auth/verify-email
// ============================================================================

// ConfirmEmailVerificationHandler is the HTTP handler for ConfirmEmailVerification.
type ConfirmEmailVerificationHandler struct {
	confirmEmailVerification application.ConfirmEmailVerification
}

// NewConfirmEmailVerificationHandler creates a new HTTP handler.
func NewConfirmEmailVerificationHandler(confirmEmailVerification application.ConfirmEmailVerification) *ConfirmEmailVerificationHandler {
	return &ConfirmEmailVerificationHandler{confirmEmailVerification: confirmEmailVerification}
}

// RegisterConfirmEmailVerificationRoute registers the HTTP route for ConfirmEmailVerification.
func RegisterConfirmEmailVerificationRoute(mux *http.ServeMux, handler *ConfirmEmailVerificationHandler) {
	mux.HandleFunc("POST /auth/verify-email", handler.ServeHTTP)
}

// Request types

// ConfirmEmailVerificationRequestBody defines the request body for ConfirmEmailVerification
type ConfirmEmailVerificationRequestBody struct {
	Token string `json:"token"`
}

// Response types

type ConfirmEmailVerificationResponse interface {
	VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error
}

type ConfirmEmailVerification200Response struct {
	Session models.Session `json:"session"`
	User    models.User    `json:"user"`
}

func (response ConfirmEmailVerification200Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type ConfirmEmailVerification400Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification400Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type ConfirmEmailVerification401Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification401Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type ConfirmEmailVerification404Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification404Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type ConfirmEmailVerification422Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification422Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type ConfirmEmailVerification429Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification429Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type ConfirmEmailVerification500Response struct {
	server.ProblemDetails
}

func (response ConfirmEmailVerification500Response) VisitConfirmEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the POST /auth/verify-email endpoint.
func (h *ConfirmEmailVerificationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := ConfirmEmailVerification401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitConfirmEmailVerificationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.ConfirmEmailVerificationInput{}
	input.SessionID = sessionID

	// Request body
	body := &ConfirmEmailVerificationRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := ConfirmEmailVerification400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitConfirmEmailVerificationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.Token = body.Token

	// Execute
	result, err := h.confirmEmailVerification.Execute(ctx, input)
	if err != nil {
		errorResp := ConfirmEmailVerification500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitConfirmEmailVerificationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := ConfirmEmailVerification200Response{}
	response.Session = result.Session
	response.User = result.User

	if err := response.VisitConfirmEmailVerificationResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}