# pkg/auth/gen/controllers/create_invitation.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// CreateInvitation - POST /organizations/{organizationID}/invitations
// ============================================================================

// CreateInvitationHandler is the HTTP handler for CreateInvitation.
type CreateInvitationHandler struct {
	createInvitation application.CreateInvitation
}

// NewCreateInvitationHandler creates a new HTTP handler.
func NewCreateInvitationHandler(createInvitation application.CreateInvitation) *CreateInvitationHandler {
	return &CreateInvitationHandler{createInvitation: createInvitation}
}

// RegisterCreateInvitationRoute registers the HTTP route for CreateInvitation.
func RegisterCreateInvitationRoute(mux *http.ServeMux, handler *CreateInvitationHandler) {
	mux.HandleFunc("POST /organizations/{organizationID}/invitations", handler.ServeHTTP)
}

// Request types

// CreateInvitationRequestBody defines the request body for CreateInvitation
type CreateInvitationRequestBody struct {
	Email string `json:"email"`
	Role  string `json:"role"`
}

// Response types

type CreateInvitationResponse interface {
	VisitCreateInvitationResponse(w http.ResponseWriter) error
}

type CreateInvitation201Response struct {
	Data models.Invitation `json:"data"`
}

func (response CreateInvitation201Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	return json.NewEncoder(w).Encode(response)
}

type CreateInvitation400Response struct {
	server.ProblemDetails
}

func (response CreateInvitation400Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateInvitation401Response struct {
	server.ProblemDetails
}

func (response CreateInvitation401Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateInvitation422Response struct {
	server.ProblemDetails
}

func (response CreateInvitation422Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateInvitation429Response struct {
	server.ProblemDetails
}

func (response CreateInvitation429Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateInvitation500Response struct {
	server.ProblemDetails
}

func (response CreateInvitation500Response) VisitCreateInvitationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the POST /organizations/{organizationID}/invitations endpoint.
func (h *CreateInvitationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := CreateInvitation401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitCreateInvitationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.CreateInvitationInput{}
	input.SessionID = sessionID

	// Path parameter "organizationID"
	var organizationID uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "organizationID", r.PathValue("organizationID"), &organizationID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := CreateInvitation400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter organizationID: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitCreateInvitationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.OrganizationID = organizationID

	// Request body
	body := &CreateInvitationRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := CreateInvitation400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitCreateInvitationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.Email = body.Email
	input.Role = body.Role

	// Execute
	result, err := h.createInvitation.Execute(ctx, input)
	if err != nil {
		errorResp := CreateInvitation500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitCreateInvitationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := CreateInvitation201Response{}
	response.Data = result.Data

	if err := response.VisitCreateInvitationResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}