# pkg/auth/gen/controllers/create_organization.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// CreateOrganization - POST /organizations
// ============================================================================

// CreateOrganizationHandler is the HTTP handler for CreateOrganization.
type CreateOrganizationHandler struct {
	createOrganization application.CreateOrganization
}

// NewCreateOrganizationHandler creates a new HTTP handler.
func NewCreateOrganizationHandler(createOrganization application.CreateOrganization) *CreateOrganizationHandler {
	return &CreateOrganizationHandler{createOrganization: createOrganization}
}

// RegisterCreateOrganizationRoute registers the HTTP route for CreateOrganization.
func RegisterCreateOrganizationRoute(mux *http.ServeMux, handler *CreateOrganizationHandler) {
	mux.HandleFunc("POST /organizations", handler.ServeHTTP)
}

// Request types

// CreateOrganizationRequestBody defines the request body for CreateOrganization
type CreateOrganizationRequestBody struct {
	BillingEmail   string    `json:"billingEmail"`
	OrganizationID uuid.UUID `json:"organizationID"`
}

// Response types

type CreateOrganizationResponse interface {
	VisitCreateOrganizationResponse(w http.ResponseWriter) error
}

type CreateOrganization201Response struct {
	Data models.Organization `json:"data"`
}

func (response CreateOrganization201Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	return json.NewEncoder(w).Encode(response)
}

type CreateOrganization400Response struct {
	server.ProblemDetails
}

func (response CreateOrganization400Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateOrganization401Response struct {
	server.ProblemDetails
}

func (response CreateOrganization401Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateOrganization422Response struct {
	server.ProblemDetails
}

func (response CreateOrganization422Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateOrganization429Response struct {
	server.ProblemDetails
}

func (response CreateOrganization429Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type CreateOrganization500Response struct {
	server.ProblemDetails
}

func (response CreateOrganization500Response) VisitCreateOrganizationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the POST /organizations endpoint.
func (h *CreateOrganizationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := CreateOrganization401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitCreateOrganizationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.CreateOrganizationInput{}
	input.SessionID = sessionID

	// Request body
	body := &CreateOrganizationRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := CreateOrganization400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitCreateOrganizationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.BillingEmail = body.BillingEmail
	input.OrganizationID = body.OrganizationID

	// Execute
	result, err := h.createOrganization.Execute(ctx, input)
	if err != nil {
		errorResp := CreateOrganization500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitCreateOrganizationResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := CreateOrganization201Response{}
	response.Data = result.Data

	if err := response.VisitCreateOrganizationResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}